Make testOrderedStore a bit less naive
This "fixes" the behavior in test_dnd where upon dropping an item,
it would "jump up" several rows.  This happened because the "before"
option in the put options always simply assigned 0.5 less than
the value of the item being dropped before - which may already be
not the first one of several items with the same order value.