Fix dnd tests a bit; issues were further under the surface.
OnDemandList was initializing query and queryOptions to {}
on the prototype, across instances.  Now it initializes in ctor, to
avoid code inadvertently modifying queries across instances.

However, because the entire create process was happening in List's
constructor, setting query in a child constructor was too late for
it to properly take effect upon render.  Therefore, that logic is now
moved to postscript, EXCEPT for the mixing in of params.

It's quite possible that the fundamental code flow of List itself
could use further thinking out to mitigate these kinds of difficulties.