Editor plugin: specify widget properties in widgetArgs column property
This removes the old behavior which simply sent a slightly-shadowed version
of the column object as the first argument to the widget constructor.
While that may appear convenient, it has potential for problems, as
several properties of the column object could become ambiguous, as they
share names with common widget members (e.g. className, label, get).