Resolve chance of entire-row-selection with CellSelection
This resolves a problem in Chrome/Safari where clicking on a
"border line" between two cells in a grid with CellSelection would
cause the entire row to be selected.
This is resolved by being more picky about what events we care about.

The only catch is that now, if you click on a borderline in
a problem browser, no action will occur.  Still, this seems less
destructive than altering the selection in a way one might not predict.
(It'd be impossible to "guess" which cell was intended to be clicked
without doing position calculations, which seems like overkill.)