Add startup lifecycle point (automatically called when possible)
This was kind of a long time coming; what put this over the edge
was an error in IE caused by OnDemandList onscroll code firing too soon
in the case of instances that are not initialized with a srcNodeRef.

This change delays calling resize and refresh (the latter of which
resulted in the OnDemandList error) until startup.  In order to preserve
the current "it just works" behavior, after postCreate we do a simple
check to see if the component is already visible; if so, we call
startup automatically.  Otherwise, it will need to be called
manually once the node is added to the DOM and/or displayed.

Tested various tests across browsers and have not seen regressions.