Add test for List's addCssRule method.
This test demonstrates an issue with the remove method hung off of
the return values of addCssRule - the remove method references
the rule's index at the time it is added to the stylesheet.  Thus,
if an earlier rule is removed before remove is called for this rule,
the wrong rule will be removed, since the CSSRuleList was since spliced.

It also demonstrates that we missed an IE < 9 compat issue -
stylesheets have no deleteRule method in IE < 9, but they have removeRule.