Fix issues with addCssRule method.
This addresses the issues exposed in the test page added in 10814a.
Namely, it adds logic to track actual stylesheet indices as rules
are removed and the stylesheet's CSSRuleList is thus spliced.

It also addresses the IE < 9 compat issue, calling removeRule if
deleteRule does not exist on the StyleSheet object.

In addition, this commit creates a separate stylesheet entirely
for these operations, in an attempt to ease debugging - earlier,
addCssRule would use whatever stylesheet was loaded last at each time
it was called.