Rename and revise behavior of data-dgrid-props
Now named data-dgrid-columns to avoid potential of confusing people into
thinking that it behaves like data-dojo-props, when it no longer does.

Previously, GridFromHtml would automatically wrap the value of this
attribute in { ... }, the same way data-dojo-props works; however, this
makes it unusable for cell plugins, so I've removed that behavior.