Refactor sort to additionally support passing a sort options array
This enables us to run criteria passed via queryOptions to OnDemandList
through the sort function of the list/grid, which in the case of Grid,
also updates the sort arrow in the UI.

It might also enable us to consider supporting nested sort via a
plugin in the future.  Haven't thought too hard about that yet.