Add RTL test support to skin.html; revert a previous border tweak
Reverting the border tweak is kind of a hard call - it sort of makes
some skins look better, but others not so much, and more importantly,
the extra rule under .dgrid-rtl could cause specificity wars with
skin styles (see: sage, cactus).  This seems like a decision better
left to the skin itself.