When editing done indented tree elements are drawn correctly.
If the editor wraps a tree, and you edit a child node, then when editing is
complete it renders with the correct level of indentation. To do this I had to
make sure that the options object used to render original child is used at the
end of the edit.