Calculate widths before actively setting sizes
This is a problem if the columns have default sizes, but the grid itself
doesn't. The tables will render out widths proportionate to the CSS requested
widths, but will typically be much larger. Then, as the code loops, the new
width on the first node causes it to be disporportionately larger than the
others, which all get crushed down quite significantly, meaning that on that
first resize, your first column gets much larger, even if you were shrinking
it.