Adjusting ColumnHider menu height dynamically
In cases with several columns and a relatively short grid, the ColumnHider
menu could exceed the bounds of the grid's viewport, but could not be
scrolled, resulting in the latter columns being inaccessible.

These changes make some small CSS modifications to allow for scrolling within
the menu if it's taller than the grid, while also adjusting its size to never
be greater than the grid itself. Given that it's currently positioned
absolutely, the height is being determined and set in code, rather than via
CSS.