Pagination: simplify _onNotification logic
This removes a special case that was initially added for if an item was
removed from the last page.  The thing is, we have no idea if that item
was in fact added to a previous page, in which case the result would be
to actually shift all preceding items in the last page down by one,
warranting a re-render of the page anyway.