fixes #886 Add ColumnHider to header node when scrollbars are hidden.
When the scrollbars are hidden on OSX we want to add the ColumnHiderIcon
node as a child of the header node. On a platform that has scrollbars we
want to add the ColumnHiderIcon as a child of the headerScrollNode.

Also fixed a small IE11 bug which was causing us to fail to add
the dgrid-rtl-swap class to the table. This would result in a small
rendering bug for positioning the headerScrollNode properly.

Finally I added a test page for the column hider extension running on a
RTL page.