Fix #886: Add ColumnHider trigger node to domNode to always be visible
dgrid 0.3.14 introduced a regression where the trigger node to open the menu
would no longer appear on browsers with no/hidden scrollbars.  This resolves it
by adding the node as a direct child of the grid's domNode instead, so it is
always visible regardless.

Added support for an rtl query parameter to the ColumnHider test page.

Thanks @treasonx for the initial investigation on this.