fixes #786 Shift + row select highlights outside text (ie)
ms-user-select: none; will block selection from starting on the element.
It will not block an existing selection from entering the element.

When using the shift modifier IE will select everything from the top of
the document, or last clicked location, up until the shift-clicked
location. Since the selection starts outside of the element marked with
ms-user-select: none; all of the text is selected.

Added a check for msUserSelect, and fallback to other methods to stop
selection.