fixes #901 CompoundColumns with sort property causes an error
Make sure that the target has contents or children before placing sort
arrow UI.

When trying to set the column sort arrow UI on a hidden column header we
were making the assumption that the column header had children. But if
the header was hidden it would cause an exception when trying to place
the UI element.