Fixes issue with dragging to end of grid
If dragging a row to the end of the grid immediately after dragging and
dropping the same row the node provided will be the drop handle, and
will result in an early return.

This commit bypasses the early return when the node provided is a drop handle
(detached from the grid) rather than an actual row.