Replace put-selector with dojo/dom-* modules
While put-selector offers terse syntax that is sometimes far easier to
write, it can also be far harder to _read_, and the amount of meaning
buried in its strings and positional arguments has had a history of
causing issues and increasing maintenance burden.

This change started partly as an experiment, but based on unit test runs,
switching to dom-construct actually improves performance up to 20%
(specifically in IE11; IE8 actually seems to benefit less).

Code size difference is negligible once minification and compression
come into play, but using dom-construct can actually still manage to be
smaller, when one considers that the dom-* modules are likely to already be
included in the majority of cases anyway.