editorArgs function for always-on editors also takes object parameter
The Editor mixin specifies a column property editorArgs, which can
either be an object or a function returning an object. If it's a
function, the function takes a column object as its first parameter.

My change was to also pass the row object as the second parameter to
editorArgs. This can only be passed for always-on editors, so for
shared editors (aka editOn editors) only the column parameter is
passed (the object parameter is undefined).

I needed to create always-on editors with parameters that depended on
the row data, and this is the only way I could see of doing it. For
editOn editors it's possible to set up a handler for the
dgrid-editor-show event and get the row data that way, but I don't
think that event is emitted for always-on editors.