Fix dgrid horizontal scroll on focus
When navigating with keyboard on a grid that is not wide enough to fit
the entire grid, ensure that the column is scrolled in to view in
entirety. Fixes issue in Chrome, Firefox, and Safari.

h/t @nicknisi for initial fix in #1086

(cherry picked from commit ecfd9a60443a88d30f13b53cf840946f03e97de5)