Fix #965: OnDemandList: Avoid loading all data if row height is 0
Also adds unit tests verifying that initially only the initial requested
range is loaded in this case, and that _processScroll will bail out if
rowHeight is 0 to avoid potential for infinite loops.

Additionally revises some _StoreMixin unit tests not to unwittingly
create grids with no columns, which doesn't make sense and is irrelevant
to what they are testing.  (Incidentally, it initially pointed out a
regression, which is why the _processScroll test exists now.)