Fix #1069: Tree: Continue to listen for transitionend events
The container only gets created once, so we need to listen for more than the
first transitionend event, in order to remove the fixed height which could
interfere with inner parent expansion.

Additionally, the immediate call to `_onTreeTransitionEnd` is unnecessary,
since the no-transition path skips height adjustment and already handles
updating the display property elsewhere.

Thanks @jbomer, @bogool, and @kriszyp for poking at this and helping to
arrive at the solution.