Fix #997: Tree: Clear expanded state of deleted store items
When a delete notification is received from a tracked store,
the corresponding entry in the `expanded` map is deleted.
This way, if an item with the same ID is added to the store again,
it will not expand automatically.

(cherry picked from commit a8a2ce90e5ec53efec67b9eb57939b65141e1e83)