Fix #1241 - set DnD property allowNested if Tree mixin is used.
Add code to dgrid/extensions/DnD to look for us of the dgrid/Tree mixin.
If Tree is being used, set `allowNested` to true which allows dojo/dnd to
properly identify the child row container as a valid drop target.

Also update the DnD_Tree test page to allow testing this.