Add SingleQuery to extensions folder, plus docs and tests
This is based on the code in the "Rendering all store data at once"
tutorial, with the following additions:

* Properly reflects total via get('total') (which will also be added to
  the tutorial)
* Emits dgrid-refresh-complete event like ODL and Pagination
* Supports loadingMessage and noDataMessage

Where possible, code has been added to _StoreMixin to be shared across
OnDemandList, Pagination, and SingleQuery.

This also makes noDataNode appear more consistently; previously,
it would not be rendered when the last row is removed if noDataMessage
isn't defined.  It also fixes #1277 which was a pre-existing bug that
became more obvious and surfaced in a test case after this change.

Thanks @msssk for narrowing down the cause of #1277 and suggesting
further refactoring.