Fix #1278: Keyboard: Programatically scroll for home/end keys
Smooth scrolling by the browser causes renderArray to be
unnecessarily called multiple times in OnDemandList when the
home or end key is pressed. This also causes the grid to display
the wrong data because an aspect on renderArray fires for the
first set of rows when it is expecting to only fire once.

Preventing the default behavior and programmatically
scrolling instead resolves this issue.

This fix also reduces the number of rows in the performance
test from 20000 to 14500. This reduced number creates a
grid that is slightly under the maximum height limit in
IE, so that home/end functionality can be tested in a
functional test in all browsers.