Keep track of focused editor cell so it persists after autosave
Editor cells using 'dgrid-cellfocusin' trigger an edit of the newly focused
cell followed by a row save (if autosave is true). The save process destroys
the row, updates the data, and re-renders the row. The result is that the
newly rendered row is no longer focused. This PR keeps track of the focused
editor cell during the edit process so that when the save triggers a re-render
the editor is correctly re-activated and focused.

Fixes #1298