Editor: fix removeRow bug
When OnDemandList is pruning rows it sometimes calls 'removeRow' on a preload
node. Editor's 'removeRow' logic assumes that the node will always be a
valid row node which causes an error when it is not. This commit adds a guard
in `Editor#removeRow` to ensure that a dgrid row object for the node exists.

Also disable the verbose logging in the Rest.html test file by default.