_StoreMixin: clamp rows.max to a minimum of 0
If the rows of a grid are removed from bottom to top then each row removal
will decrement rows.max, including the final row removal which will set
rows.max to -1. This breaks the logic in _StoreMixin's 'add, update' event
handler preventing insertion of new rows when items are added to the store.

Fixes #1305